flask-template/object-detection-flask-app/requirements.txt
flask
tensorflow
numpy
opencv-python
Pillow
scikit-learn
flask-cors